Grab variables from several censuses on a common geography. Requires sf package to be available Will return CT level data

Grab variables from several censuses on a common geography. Requires sf package to be available Will return CT level data

get_tongfen_ca_ct(
  regions,
  vectors,
  geo_base = NULL,
  match_common_GeoUID = TRUE,
  area_mismatch_cutoff = 0.1,
  tolerance = 500,
  na.rm = FALSE,
  quiet = FALSE,
  refresh = FALSE
)

Arguments

regions
census region list, should be inclusive list of GeoUIDs across censuses

vectors
List of cancensus vectors, can come from different census years

geo_base
optional parameter with allowed values one of the geo_datasets to base the geographic data on,

match_common_GeoUID
logical, if `TRUE` (default) assumes CTs with same GeoUID are identical

area_mismatch_cutoff
cutoff to disregard tracts that can't be matched within threshold

tolerance
tolerance in metres when matching changing CT geographies.

na.rm
remove NA values when aggregating up values, default is `TRUE`

quiet
suppress download progress output, default is `FALSE`

refresh
optional character, refresh data cache for this call for no geographic data, (default `NA`) or `NULL` (the default) for not returning geographic data`

Value

dataframe with census variables on common geography