A dataset with polling district geographies from the 2015 federal election in the Vancouver area

A dataset with polling district geographies from the 2015 federal election in the Vancouver area

References

https://www.elections.ca/content.aspx?section=res&dir=rep/off&document=index&lang=e#42GE

Author

Elections Canada